Moved pagination, sort. #338
